move apps_balance function from sp_updater module to apps_updater
